ISPN-5061 Cross site state transfer - NPE on consumer site when backup cache is local
* Created BackupReceiver implementation to handle the local-mode caches
